Implement a proc-macro to generate new functions
A working proc-macro for declaring custom function extensions was added
in this commit. It still is a work in progress, so I am checking in here
now that it at least works. It still requires a lot of work to handle
correctness in the function definition, and also will probably change as
the WG sorts out the type system for functions.