Restructure crate for macros
This commit reorganizes the crate so that it can support proc macros.
The primary crate was moved down a level into `serde_json_path` and a
new crate, `serde_json_path_macros` was created alongside it.
`serde_json_path` is still the primary crate, and all of its main
components were moved down, including:

- src/ folder
- test/ folder
- Cargo.toml
- LICENSE
- README
- CHANGELOG

One exception was that some of the build flags in the Cargo.toml file
had to be elevated up to the workspace Cargo.toml file.

The new `serde_json_path_macros` crate has its own unique structure. It
contains an `internal` crate of its own. This is where the actual proc
macro is defined. This was done because the proc macro we are defining
needs to rely on external dependencies; namely, `inventory` and
`once_cell`, and we don't want to put the burden on the user to have to
import these crates into their own libraries/applications.