Get things in working order
The parsing errors were re-worked to actually propagate to the top. This
required writing impls for any external error type generated by a
map_res throughout the code.

The tests for functions were removed from the spec_examples tests into
the functions tests.

Some parser alt arms were re-ordered, and the function parser was
wrapped in cut to cause failure when function validation fails. Not sure
if this is the correct solution yet, but everything is passing for now.