Split Registry from User-defined Functions
This solves the problem of having potential inventory collisions between
standard registry functions and user-defined functions. Now, when
validating queries, the inventory of user-defined functions is checked
first, and the function validated if there is a hit. If there is no hit,
then the REGISTRY is checked for one of the standard defined functions.

This ensures that if users define functions with names that later become
standardized (and added to serde_json_path), then their functions will
still take priority.

The feature gating was changed so that the registry, which is part of
the standard, is always built-in to serde_json_path, while the ability
to define custom functions is now behind the "functions" feature flag.
For now, this feature is enabled by default.

To do this, a new attribute macro, `#[register]`, was added. This is
purely for internal use within the serde_json_path crate, and serves no
purpose for external users.

Another benefit worth noting, is that the standard registry functions
are now built-in using a HashMap, which will ensure fast lookups.